Initial healthbar implementation
Need to get at the bar from the enemyAI script. Started code, commented it out
Currently using hack (script on the bar itself) which won't work when we have multiple enemies, but proves that the bar can be shortened. The Anchor is a small invisible cube that I hoped to use as a reference point so that we keep the health bar left-aligned. Currently, it's centered