Added a New Character WITH ANIMATION.
The old one is still a prefab, so I deleted him in every scene and put in the new one.

The new one didn't have correctly functioning meshes, so he currently has a skin tone on all areas of his body. We can hide this by giving him a default set of armor.

 I had some issues getting more than one animation to work previously, but now that I have the cameras working properly, I can try to mess with it again. But Currently, running is implemented and that's really the important one.